Unifies the decode macro for both Scrooge and Humbug.
